Move assert orders to own package
