+++ trunk/source/Library-TraceCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/tracecheckerutils/assertorders/AssertOrderInsideLoopFirst1.java
@@ -1,4 +1,4 @@
-package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck;
+package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.assertorders;
 
 import java.util.List;
 import java.util.Map;
@@ -9,12 +9,12 @@ import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.I
 import de.uni_freiburg.informatik.ultimate.util.datastructures.relation.HashTreeRelation;
 
 /**
- * Assert statements in incremental order by their depth, and check after each step for satisfiability. E.g. first
- * assert all statements with depth 0, then assert all statements at depth 1, and so on.
+ * Assert statements in decremental order by their depth, and check after each step for satisfiability. E.g. first
+ * assert all statements with depth max_depth, then assert all statements of depth max_depth - 1, and so on.
  *
  * @author musab@informatik.uni-freiburg.de
  */
-public class AssertOrderOutsideLoopFirst2<L extends IAction> extends AssertOrder<L> {
+public class AssertOrderInsideLoopFirst1<L extends IAction> extends AssertOrder<L> {
 	@Override
 	public List<Set<Integer>> partitionTrace(final NestedWord<L> trace,
 			final List<Object> controlConfigurationSequence) {
@@ -22,6 +22,7 @@ public class AssertOrderOutsideLoopFirst2<L extends IAction> extends AssertOrder
 				computeRelationWithTreeSetForTrace(0, trace.length(), controlConfigurationSequence);
 		final Map<Integer, Set<Integer>> depth2Statements =
 				partitionStatementsAccordingDepth(trace, rwt, controlConfigurationSequence);
-		return depth2Statements.keySet().stream().sorted().map(depth2Statements::get).toList();
+		return depth2Statements.keySet().stream().sorted((i1, i2) -> i2.compareTo(i1)).map(depth2Statements::get)
+				.toList();
 	}
 }
