+++ trunk/source/Library-TraceCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/tracecheckerutils/assertorders/AssertOrderMixInsideOutside.java
@@ -1,4 +1,4 @@
-package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck;
+package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.assertorders;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
