+++ trunk/source/Library-TraceCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/tracecheckerutils/assertorders/AssertOrderOutsideLoopFirst1.java
@@ -1,4 +1,4 @@
-package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck;
+package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.assertorders;
 
 import java.util.List;
 import java.util.Map;
